+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### 🏡 Features
+- Add User types history resource items ([#412](https://github.com/whirli/whirli-client/pull/412))
+
+## [0.0.85] - 2022-02-21
+### 🏡 Features
+- Add shipping manifest to return order ([#413](https://github.com/whirli/whirli-client/pull/413))
 
 ## [0.0.84] - 2022-02-21
 ### 🏡 Features